refactor: streamline skill documentation and enhance SVG components for better clarity and performance
