+++ .opencode/skills/cleanup/SKILL.md
@@ -3,45 +3,11 @@ name: cleanup
 description: Code cleanup specialist
 ---
 
-# Cleanup Audit
+Return a numeric list of cleanup options:
 
-Audit implementations, callers, and tests for the smallest behavior-preserving improvements. Inspect enough of each flow to distinguish redundant work from behavior required for correctness.
-
-Do not edit files unless the user explicitly asks. Do not recommend speculative rewrites, broad refactors, or memoization without demonstrated value.
-
-## Backend Focus
-
+- Follow the `Python Guidelines` section. If anything does not follow the guidelines, propose a fix. Else, move on and check the next steps.
 - Find redundant queries, refreshes, reloads, eager loads, N+1 calls, transformations, validation, parameters, return values, wrappers, and abstractions.
-- Prefer returning raw ORM objects, domain objects, dictionaries, lists, or primitives when declared response schemas already provide filtering and serialization.
-- Preserve required relationship loading, transaction boundaries, row locks, isolation, idempotency, race protections, and concurrency guarantees.
-- Trace route, service, database, and schema boundaries before treating validation or loading as redundant.
-
-## Frontend Focus
-
-- Audit components, hooks, routes, API clients, state management, callers, and tests.
-- Find redundant requests, refetches, cache invalidations, effects, renders, local state, derived state, transformations, validation, props, return values, wrappers, and abstractions.
-- Identify duplicated loading, error, pagination, form, and URL-state logic, request waterfalls, and client-side N+1 patterns.
-- Prefer rendering typed API or domain data directly instead of copying it into local state or reshaping it unnecessarily.
-- Preserve cache correctness, request cancellation, race protections, authentication boundaries, accessibility, responsive behavior, user interactions, and loading and error states.
-- Do not recommend memoization unless an existing performance problem or expensive recomputation demonstrates the need.
-
-## Method
-
-1. Trace the implementation through its callers, data boundaries, and relevant tests.
-2. Verify whether each apparent redundancy protects observable behavior or correctness.
-3. Prefer deleting work or inlining a thin abstraction over introducing new helpers or architecture.
-4. Recommend the narrowest change that preserves current behavior.
-5. Include only concrete, actionable findings. Do not invent findings to fill a quota.
-
-## Output
-
-Return numbered findings ordered by expected value. For each finding include:
-
-- File and line links for the implementation and important callers.
-- The redundant or unnecessary work and why it is safe to remove or simplify.
-- The minimal change.
-- The expected benefit.
-- The implementation risk.
-- The affected existing tests, including test file links, or state that no relevant tests were found.
-
-If no worthwhile simplifications are found, say so explicitly and note any areas that could not be verified.
+- Find dead code, unused imports, unused variables, unused parameters, unused return values, and unused tests.
+- Check for outdated libraries, or unused dependencies, and propose updates or removals.
+- Simplify the test cases, remove unnecessary mocks, and reduce test duplication.
+- Check the development workflow and the repository.
