+++ web/src/svg/FastAPI.tsx
@@ -1,0 +1,13 @@
+import type { SVGProps } from 'react';
+
+/** Renders the FastAPI mark. */
+const FastAPI = (props: SVGProps<SVGSVGElement>) => (
+    <svg {...props} preserveAspectRatio="xMidYMid" viewBox="0 0 256 256">
+        <path
+            d="M128 0C57.33 0 0 57.33 0 128s57.33 128 128 128 128-57.33 128-128S198.67 0 128 0Zm-6.67 230.605v-80.288H76.699l64.128-124.922v80.288h42.966L121.33 230.605Z"
+            fill="#009688"
+        />
+    </svg>
+);
+
+export { FastAPI };
